+++ sleap/gui/utils.py
@@ -1,6 +1,7 @@
 """Generic module containing utilities used for the GUI."""
 
 import zmq
+import time
 from typing import Optional
 
 
@@ -12,6 +13,7 @@ def is_port_free(port: int, zmq_context: Optional[zmq.Context] = None) -> bool:
     try:
         socket.bind(address)
         socket.unbind(address)
+        time.sleep(0.5)
         return True
     except zmq.error.ZMQError:
         return False
@@ -25,4 +27,29 @@ def select_zmq_port(zmq_context: Optional[zmq.Context] = None) -> int:
     socket = ctx.socket(zmq.REP)
     port = socket.bind_to_random_port("tcp://127.0.0.1")
     socket.close()
+    return port
+
+
+def find_free_port(port: int, zmq_context: zmq.Context):
+    """Find free port to bind to.
+
+    Args:
+        port: The port to start searching from.
+        zmq_context: The ZMQ context to use.
+
+    Returns:
+        The free port.
+    """
+    attempts = 0
+    max_attempts = 10
+    while not is_port_free(port=port, zmq_context=zmq_context):
+        if attempts >= max_attempts:
+            raise RuntimeError(
+                f"Could not find free port to display training progress after "
+                f"{max_attempts} attempts. Please check your network settings "
+                "or use the CLI `sleap-train` command."
+            )
+        port = select_zmq_port(zmq_context=zmq_context)
+        attempts += 1
+
     return port
