+++ sleap/gui/widgets/monitor.py
@@ -13,7 +13,7 @@ from matplotlib.collections import PathCollection
 import matplotlib.transforms as mtransforms
 from qtpy import QtCore, QtWidgets
 
-from sleap.gui.utils import is_port_free, select_zmq_port
+from sleap.gui.utils import find_free_port
 from sleap.gui.widgets.mpl import MplCanvas
 from sleap.nn.config.training_job import TrainingJobConfig
 
@@ -785,30 +785,6 @@ class LossViewer(QtWidgets.QMainWindow):
         self.ctx_given = zmq_context is not None
         self.ctx = zmq.Context() if zmq_context is None else zmq_context
 
-        def find_free_port(port: int, zmq_context: zmq.Context):
-            """Find free port to bind to.
-
-            Args:
-                port: The port to start searching from.
-                zmq_context: The ZMQ context to use.
-
-            Returns:
-                The free port.
-            """
-            attempts = 0
-            max_attempts = 10
-            while not is_port_free(port=port, zmq_context=zmq_context):
-                if attempts >= max_attempts:
-                    raise RuntimeError(
-                        f"Could not find free port to display training progress after "
-                        f"{max_attempts} attempts. Please check your network settings "
-                        "or use the CLI `sleap-train` command."
-                    )
-                port = select_zmq_port(zmq_context=self.ctx)
-                attempts += 1
-
-            return port
-
         # Progress monitoring, SUBSCRIBER
         self.sub = self.ctx.socket(zmq.SUB)
         self.sub.subscribe("")
@@ -818,9 +794,6 @@ class LossViewer(QtWidgets.QMainWindow):
             port=self.zmq_ports["publish_port"], zmq_context=self.ctx
         )
         publish_address = f"tcp://127.0.0.1:{self.zmq_ports['publish_port']}"
-
-        # Wait a bit to make sure the socket is ready.
-        time.sleep(0.5)
         self.sub.bind(publish_address)
 
         # Controller, PUBLISHER
@@ -1118,12 +1091,14 @@ class LossViewer(QtWidgets.QMainWindow):
         if self.sub is not None:
             self.sub.unbind(self.sub.LAST_ENDPOINT)
             self.sub.close()
+            time.sleep(0.1)
             self.sub = None
 
         if self.zmq_ctrl is not None:
             url = self.zmq_ctrl.LAST_ENDPOINT
             self.zmq_ctrl.unbind(url)
             self.zmq_ctrl.close()
+            time.sleep(0.1)
             self.zmq_ctrl = None
 
         # If we started out own zmq context, terminate it.
